requirements.txt
Django==1.11.5
PyJWT==1.5.3
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
django-braces==1.11.0
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.22
whitenoise==3.3.1
psycopg2==2.6.2
